\name{ABC}
\alias{ABC}
\docType{data}
\title{
ABC
}
\description{
A customer satisfaction survey, ABC.csv. The data consists of 266 responses to a questionnaire with a question on overall satisfaction (q1) and responses to 125 other questions. Variables named as "q" and a number are satisfaction or agreement levels on a 0-5 scale, Variables named as "qi" represent importance on a 1-3 scale.
}
\usage{data("ABC")}
\format{
  A data frame with 266 observations on the following 134 variables.
  \describe{
    \item{\code{q1}}{a numeric vector}
    \item{\code{q4}}{a numeric vector}
    \item{\code{q5}}{a numeric vector}
    \item{\code{q6}}{a numeric vector}
    \item{\code{qi6}}{a numeric vector}
    \item{\code{q7}}{a numeric vector}
    \item{\code{qi7}}{a numeric vector}
    \item{\code{q8}}{a numeric vector}
    \item{\code{qi8}}{a numeric vector}
    \item{\code{q9}}{a numeric vector}
    \item{\code{qi9}}{a numeric vector}
    \item{\code{q10}}{a numeric vector}
    \item{\code{qi10}}{a numeric vector}
    \item{\code{q11}}{a numeric vector}
    \item{\code{q12}}{a numeric vector}
    \item{\code{qi12}}{a numeric vector}
    \item{\code{q13}}{a numeric vector}
    \item{\code{qi13}}{a numeric vector}
    \item{\code{q14}}{a numeric vector}
    \item{\code{qi14}}{a numeric vector}
    \item{\code{q15}}{a numeric vector}
    \item{\code{qi15}}{a numeric vector}
    \item{\code{q16}}{a numeric vector}
    \item{\code{qi16}}{a numeric vector}
    \item{\code{q17}}{a numeric vector}
    \item{\code{q18}}{a numeric vector}
    \item{\code{qi18}}{a numeric vector}
    \item{\code{q19}}{a numeric vector}
    \item{\code{qi19}}{a numeric vector}
    \item{\code{q20}}{a numeric vector}
    \item{\code{qi20}}{a numeric vector}
    \item{\code{q21}}{a numeric vector}
    \item{\code{qi21}}{a numeric vector}
    \item{\code{q22}}{a numeric vector}
    \item{\code{qi22}}{a numeric vector}
    \item{\code{q23}}{a numeric vector}
    \item{\code{qi23}}{a numeric vector}
    \item{\code{q24}}{a numeric vector}
    \item{\code{qi24}}{a numeric vector}
    \item{\code{q25}}{a numeric vector}
    \item{\code{q26}}{a numeric vector}
    \item{\code{qi26}}{a numeric vector}
    \item{\code{q27}}{a numeric vector}
    \item{\code{qi27}}{a numeric vector}
    \item{\code{q28}}{a numeric vector}
    \item{\code{qi28}}{a numeric vector}
    \item{\code{q29}}{a numeric vector}
    \item{\code{qi29}}{a numeric vector}
    \item{\code{q30}}{a numeric vector}
    \item{\code{qi30}}{a numeric vector}
    \item{\code{q31}}{a numeric vector}
    \item{\code{q32}}{a numeric vector}
    \item{\code{qi32}}{a numeric vector}
    \item{\code{q33}}{a numeric vector}
    \item{\code{qi33}}{a numeric vector}
    \item{\code{q34}}{a numeric vector}
    \item{\code{qi34}}{a numeric vector}
    \item{\code{q35}}{a numeric vector}
    \item{\code{qi35}}{a numeric vector}
    \item{\code{q36}}{a numeric vector}
    \item{\code{qi36}}{a numeric vector}
    \item{\code{q37}}{a numeric vector}
    \item{\code{qi37}}{a numeric vector}
    \item{\code{q38}}{a numeric vector}
    \item{\code{q39}}{a numeric vector}
    \item{\code{q40}}{a numeric vector}
    \item{\code{q41}}{a numeric vector}
    \item{\code{qi41}}{a numeric vector}
    \item{\code{q42}}{a numeric vector}
    \item{\code{q43}}{a numeric vector}
    \item{\code{q44}}{a numeric vector}
    \item{\code{q45}}{a numeric vector}
    \item{\code{qi45}}{a numeric vector}
    \item{\code{q46}}{a numeric vector}
    \item{\code{qi46}}{a numeric vector}
    \item{\code{q47}}{a numeric vector}
    \item{\code{qi47}}{a numeric vector}
    \item{\code{q48}}{a numeric vector}
    \item{\code{qi48}}{a numeric vector}
    \item{\code{q49}}{a numeric vector}
    \item{\code{q50}}{a numeric vector}
    \item{\code{qi50}}{a numeric vector}
    \item{\code{q51}}{a numeric vector}
    \item{\code{qi51}}{a numeric vector}
    \item{\code{q52}}{a numeric vector}
    \item{\code{qi52}}{a numeric vector}
    \item{\code{q53}}{a numeric vector}
    \item{\code{qi53}}{a numeric vector}
    \item{\code{q54}}{a numeric vector}
    \item{\code{qi54}}{a numeric vector}
    \item{\code{q55}}{a numeric vector}
    \item{\code{qi55}}{a numeric vector}
    \item{\code{q56}}{a numeric vector}
    \item{\code{qi56}}{a numeric vector}
    \item{\code{q57}}{a numeric vector}
    \item{\code{q58}}{a numeric vector}
    \item{\code{qi58}}{a numeric vector}
    \item{\code{q59}}{a numeric vector}
    \item{\code{qi59}}{a numeric vector}
    \item{\code{q60}}{a numeric vector}
    \item{\code{qi60}}{a numeric vector}
    \item{\code{q61}}{a numeric vector}
    \item{\code{qi61}}{a numeric vector}
    \item{\code{q62}}{a numeric vector}
    \item{\code{qi62}}{a numeric vector}
    \item{\code{q63}}{a numeric vector}
    \item{\code{qi63}}{a numeric vector}
    \item{\code{q64}}{a numeric vector}
    \item{\code{qi64}}{a numeric vector}
    \item{\code{q65}}{a numeric vector}
    \item{\code{q66}}{a numeric vector}
    \item{\code{qi66}}{a numeric vector}
    \item{\code{q67}}{a numeric vector}
    \item{\code{q68}}{a numeric vector}
    \item{\code{q70}}{a numeric vector}
    \item{\code{q71}}{a numeric vector}
    \item{\code{q72}}{a numeric vector}
    \item{\code{q73}}{a numeric vector}
    \item{\code{q74}}{a numeric vector}
    \item{\code{q75}}{a numeric vector}
    \item{\code{q76}}{a numeric vector}
    \item{\code{q77}}{a numeric vector}
    \item{\code{q78}}{a numeric vector}
    \item{\code{q79}}{a numeric vector}
    \item{\code{q80}}{a numeric vector}
    \item{\code{q81}}{a numeric vector}
    \item{\code{var1}}{Continent, a factor with levels \code{Europe}}
    \item{\code{var3}}{Country, a factor with levels \code{Benelux} \code{France} \code{Germany} \code{Italy} \code{UK} \code{Israel}}
    \item{\code{var4}}{Segmentation, a factor with levels \code{Other} \code{Silver} \code{Gold} \code{Platinum}}
    \item{\code{var6}}{Age of ABC's equipment, a factor with levels \code{less than 1} \code{1-2} \code{2-3} \code{3-4} \code{more than 4}}
    \item{\code{var9}}{Profitability, a factor with levels \code{Profitable} \code{Break-Even} \code{Below Break-Even}}
    \item{\code{var11}}{Position, a factor with levels \code{Owner} \code{Management} \code{Technical management} \code{Technical staff} \code{Operator} \code{Administrator} \code{Other}}
    \item{\code{customerSeniority}}{a numeric vector}
    \item{\code{country}}{a factor with levels \code{Benelux} \code{France} \code{Germany} \code{Israel} \code{Italy} \code{UK}}
  }
}
\source{
Kenett, R and Salini, S. (2013)
\emph{Modern Analysis of Customer Surveys}.
}
\references{
The ABC Company, Questionnaire and Data
}
\examples{
data(ABC)
barplot(table(ABC$q1, ABC$q4))
}
\keyword{datasets}
