\name{OELECT}
\alias{OELECT}
\docType{data}
\title{
Electric Voltage Outputs of Rectifying Circuits
}
\description{
99 electric voltage outputs of a rectifying circuit (\eqn{V}).
}
\usage{data(OELECT)}
\source{
  Kenett, R. and Zacks, S. (1998)
  \emph{Modern Industrial Statistics: The Design and Control of Quality and Reliability}.
  Duxbury Press.
}
\examples{
data(OELECT)

summary(OELECT)

mean(OELECT)
}
\keyword{datasets}
