\name{DOW1941}
\alias{DOW1941}
\docType{data}
\title{
Dow-Jones Financial Index 1941
}
\description{
The Dow-Jones daily index of 1941.
}
\usage{data(DOW1941)}
\source{
  Kenett, R. and Zacks, S. (1998)
  \emph{Modern Industrial Statistics: The Design and Control of Quality and Reliability}.
  Duxbury Press.
}
\examples{
data(DOW1941)

plot(DOW1941,          
     type="b", 
     ylab="Dow Jones 1941")
}
\keyword{datasets}
